+++ packages/orq-rc/src/orq_ai_sdk/models/bulkfileuploadop.py
@@ -19,7 +19,7 @@ class BulkFileUploadFilesTypedDict(TypedDict):
 
 class BulkFileUploadFiles(BaseModel):
     file_name: Annotated[
-        str, pydantic.Field(alias="files"), FieldMetadata(multipart=True)
+        str, pydantic.Field(alias="fileName"), FieldMetadata(multipart=True)
     ]
 
     content: Annotated[
@@ -81,5 +81,5 @@ class ResponseBody(BaseModel):
 
     file_name: str
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2024-12-01T21:30:45.101Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2024-12-16T16:28:54.131Z")
     r"""The date and time the resource was created"""